zram-swap: modify the size configuration parameters
Tweak the configuration to account for the expected compression ratio, when
defining the zram device size. For this purpose, we introduce two new
parameters, preset to conservative defaults.

zram_ram_percent: maximum percentage of the total memory which the zram device
is allowed to use. Self-explanatory, defaults to 10.

zram_exp_ratio: expected compression ratio of the zram device. This depends on
the selected compression algorithm. Both lzo and lz4 have a compression ratio
of around 2(:1), lzo-rle around 3.7, and zstd around 4.5. The default algorithm
is lzo, so this value defaults to 2.

This change also obsoletes the previous zram_size_mb parameter, too coarse to
tune the zram size on embedded devices with small amounts of memory.

Signed-off-by: Rui Salvaterra <rsalvaterra@gmail.com>